qwen4_exp: cuBLASLt for the decode-shaped hc collapse — decode +22-25%
The lever-2 stage profile put the mHC bracket at ~19 ms of every ~65 ms
decode token: hc_pre_attn 254 us + hc_post/pre_ffn 265 us per GDN layer,
against a ~150-200 us bandwidth floor. The small-T split path's k_down/
k_fin each stream ~6.5 MB of low-rank weights through hand-rolled
kernels — the same GEMM-shaped-work-on-hand-rolled-kernels defect class
as the prefill collapse (925a1b5e) and the batched-decode QKVZ arms
(8da3fb22), and the same cure.

Small T (<=64) now routes through the EXISTING GEMM decomposition
(stage_bf16 -> down -> silu -> up -> inject -> mix) with cuBLASLt for
the three projections; the prefill regime keeps dense_gemm_bf16
_pipelined; ATLAS_HC_DECODE_SPLIT=1 keeps the split arm (A/B). The
hc head (model-level mixer, once per token) rides the same switch.

MEASURED at matched SM clocks (2281 MHz, ssm/moe/ple stages as the
control — unchanged to the microsecond):
  hc_pre_attn        254 -> 122 us/layer
  hc_post+hc_pre_ffn 265 -> 131 us/layer   (~9.6 ms/token total)
  decode C=1         61.5-68 -> 48.9-55.9 ms median (16.3 -> 17.9-20.5 tok/s)
  C=2  21.8 -> 25.3 aggregate   C=4  30.1 -> 32.6 aggregate
Greedy text byte-identical; C=2 content + QSA retrieval probes clean.

Parity: the cuBLASLt arm's reduction order differs from the tile GEMM,
worst element ~9% of ref RMS at cos >= 0.999995 (the grade the prefill
GEMM shipped with); the fixture now tests BOTH small-T arms — the split
arm holds the original 5% bound, the cublas arm 12% + the cosine gate.
hc_pre_gemm split to hyper_connection_lowrank_gemm.rs (500-LoC cap).

Gates: fmt, clippy --all-targets, LoC cap, 2/2 hc GPU parity tests.
